Agent_Templates/Runtime_envs/Cloud_Run/app/orchestration/requirements.txt
google-cloud-aiplatform==1.44.0
google-cloud-bigquery==3.21.0
google-cloud-discoveryengine==0.11.10
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.42
langchain-experimental==0.0.34
langchain-google-genai==0.0.11
langchain-google-vertexai==0.1.3
pandas==2.0.3
pydantic==2.4.2
vertexai==1.47.0